% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_report.R
\name{check_and_install}
\alias{check_and_install}
\title{Check and Install}
\usage{
check_and_install(packs)
}
\arguments{
\item{packs}{Vector of package names}
}
\value{
data.frame, status of required packages and their installation status
}
\description{
Check whether given packages are installed or not and if not installed install them
}
\details{
check_and_install
}
\examples{
check_and_install(c('dplyr', 'data.table'))
}
