% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_matrix.R
\name{confmatrix}
\alias{confmatrix}
\title{confmatrix}
\usage{
confmatrix(actuals, preds)
}
\arguments{
\item{actuals}{factor}

\item{preds}{factor}
}
\value{
table
}
\description{
confmatrix
}
\examples{
\dontrun{
confmatrix(c(1,1,1,0), c(1,1,0,0))
}
}
\author{
Jayachandra N
}
