% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_partition.R
\name{dataPartition}
\alias{dataPartition}
\title{dataPartition}
\usage{
dataPartition(df, train_data_perc)
}
\arguments{
\item{df}{data.frame which need to be devided into train and test subset}

\item{train_data_perc}{numeric value between 1 to 100}
}
\value{
list of length 2 which contains Train data and Test data
}
\description{
dataPartition
}
\examples{
\dontrun{
dataPartition(iris, 80)
}
}
\author{
Jayachandra N
}
