% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_these.R
\name{excludeThese}
\alias{excludeThese}
\title{excludeThese}
\usage{
excludeThese(set, items_to_exclude)
}
\arguments{
\item{set}{vector}

\item{items_to_exclude}{vector to exclude from the whole set}
}
\value{
vector
}
\description{
excludeThese
}
\examples{
\dontrun{
excludeThese(1:10, 1)
}
}
\author{
Jayachandra N
}
