% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numeric_cols.R
\name{getcharacterCols}
\alias{getcharacterCols}
\title{getcharacterCols}
\usage{
getcharacterCols(dat)
}
\arguments{
\item{dat}{data frame}
}
\value{
A Character vector of names of numeric columns of a given data frame
}
\description{
getcharacterCols
}
\author{
Jayachandra N
}
