% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefficients.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Get Coefficients}
\usage{
getCoefficients(model)
}
\arguments{
\item{model}{lm model}
}
\value{
data.frame of coeffcients
}
\description{
Get coefficients from the model summary
}
\details{
getCoefficients
}
\examples{
 model <- lm(Sepal.Length ~ ., iris) # A linear regression model
 getCoefficients(model)
}
\author{
Jayachandra N
}
