% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_insights.R
\name{getDataInsight}
\alias{getDataInsight}
\title{getDataInsight}
\usage{
getDataInsight(temp)
}
\arguments{
\item{temp}{data frame}
}
\value{
list of details of data
}
\description{
getDataInsight
}
\examples{
getDataInsight(mtcars)
getDataInsight(iris)
}
\author{
Jayachandra N
}
