% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_report.R
\name{getLibraryReport}
\alias{getLibraryReport}
\title{getLibraryReport}
\usage{
getLibraryReport(packages)
}
\arguments{
\item{packages}{Vector of package names}
}
\value{
data.frame, status of required packages and their installation status
}
\description{
getLibraryReport
}
\examples{
check_and_install(c('dplyr', 'data.table'))
}
