% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_most_repeated_value.R
\name{getMostRepeatedValue}
\alias{getMostRepeatedValue}
\title{getMostRepeatedValue}
\usage{
getMostRepeatedValue(vec)
}
\arguments{
\item{vec}{Vector to calculate most repeated values}
}
\value{
most repeated values in the given set of values
}
\description{
getMostRepeatedValue
}
\examples{
\dontrun{
getMostRepeatedValue(c(1,2,3,3,3,2))
getMostRepeatedValue(c("R", "R", "Python", "Python", "R"))
}
}
