% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numeric_cols.R
\name{getnumericCols}
\alias{getnumericCols}
\title{Get Numeric Cols}
\usage{
getnumericCols(dat)
}
\arguments{
\item{dat}{data frame}
}
\value{
Character vector of names of numeric columns of given data frame
}
\description{
Get all columns which are numeric.
}
\details{
getnumericCols
}
\examples{
getnumericCols(iris)
getnumericCols(mtcars)
}
\author{
Jayachandra N
}
