% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type.R
\name{getTypeOfColumns}
\alias{getTypeOfColumns}
\title{getTypeOfColumns}
\usage{
getTypeOfColumns(df)
}
\arguments{
\item{df}{data frame}
}
\value{
Data frame of column name and it's type
}
\description{
getTypeOfColumns
}
\examples{
getTypeOfColumn(mtcars)
getTypeOfColumn(iris)
}
\author{
Jayachandra N
}
