% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cloud.R
\name{getWordCloud}
\alias{getWordCloud}
\title{Get Word Cloud}
\usage{
getWordCloud(d)
}
\arguments{
\item{d}{table of word's frequency}
}
\value{
Word cloud plot
}
\description{
Get word cloud for given table of words' frequencies
}
\details{
getWordCloud
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- getFeqTable("Hello! R is Great")
getWordCloud(x)
\dontshow{\}) # examplesIf}
}
