% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_var.R
\name{make_var}
\alias{make_var}
\title{make_var}
\usage{
make_var(prefix, var, suffix)
}
\arguments{
\item{prefix}{prefix character}

\item{var}{character to convert}

\item{suffix}{suffix character}
}
\value{
variable
}
\description{
make_var
}
\examples{
\dontrun{
make_var("", "Jay", "")
make_var("", "Incredible_India", "")
}
}
