% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{multinomial}
\alias{multinomial}
\title{multinomial}
\usage{
multinomial(eqn, df)
}
\arguments{
\item{eqn}{formula to build model}

\item{df}{data frame}
}
\value{
model
}
\description{
multinomial
}
\examples{
\dontrun{
multinomial( Species ~ .,  iris)
}
}
