% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyR.R
\name{plotCor}
\alias{plotCor}
\title{plotCor}
\usage{
plotCor(cor_dat, my_method)
}
\arguments{
\item{cor_dat}{Corelation matrix}

\item{my_method}{method to plot Example: circle}
}
\value{
Corelation plot
}
\description{
plotCor
}
