% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_model.R
\name{randomForestModel}
\alias{randomForestModel}
\title{randoMForestModel}
\usage{
randomForestModel(eqn, df)
}
\arguments{
\item{eqn}{formula}

\item{df}{data.frame}
}
\value{
rf model
}
\description{
randoMForestModel
}
\examples{
\dontrun{
#' mod <- randomForestModel( Species ~ .,  iris)
}
}
