% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyR.R
\name{shineMe}
\alias{shineMe}
\title{shineMe}
\usage{
shineMe()
}
\value{
shiny UI page
}
\description{
An R shiny app for shinyr UI.
}
\details{
shineMe
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
shineMe()
\dontshow{\}) # examplesIf}
}
\author{
Jayachandra N
}
