% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_and_get.R
\name{splitAndGet}
\alias{splitAndGet}
\title{splitAndGet}
\usage{
splitAndGet(x)
}
\arguments{
\item{x}{string to split into words}
}
\value{
List of worrds
}
\description{
splitAndGet
}
\examples{
\dontrun{
splitAndGet("R programming is awesome!")
}
}
\author{
Jayachandra N
}
