% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{valid_sets}
\alias{valid_sets}
\title{valid_sets}
\usage{
valid_sets(package = NULL, cols = NULL)
}
\arguments{
\item{package}{package name to fetch inbuilt data sets example: "datasets"}

\item{cols}{numeric to specify condition on how many columns should data frame have}
}
\value{
data frame all available datasets of class data frame
}
\description{
valid_sets
}
\examples{
\dontrun{
valid_sets()
}
}
\author{
Ravindra Pushker

Jayachandra N
}
