% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_report.R
\name{check_and_install}
\alias{check_and_install}
\title{check_and_install}
\usage{
check_and_install(packs)
}
\arguments{
\item{packs}{Vector of package names}
}
\value{
data.frame, status of required packages and their installation status
}
\description{
Tries to install and load the packages which are not installed and returns the status
}
\examples{
check_and_install(c('dplyr', 'data.table'))
}
