% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_class.R
\name{detectClass}
\alias{detectClass}
\title{detectClass}
\usage{
detectClass(x)
}
\arguments{
\item{x}{a vector}
}
\value{
type of the vector
}
\description{
detectClass
}
\examples{
\dontrun{
detectClass(c(1,2,3))
detectClass(c("a","b"))
detectClass(iris$Species)
}
}
\author{
Jayachandra N
}
