% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefficients.R
\name{getCoefficients}
\alias{getCoefficients}
\title{getCoefficients}
\usage{
getCoefficients(model)
}
\arguments{
\item{model}{lm model}
}
\value{
data.frame of coeffcients
}
\description{
getCoefficients
}
\examples{
\dontrun{
 x <- lm(Sepal.Length ~ ., iris)
 getCoefficients(x)
}
}
\author{
Jayachandra N
}
