% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freq_table.R
\name{getFeqTable}
\alias{getFeqTable}
\title{getFeqTable}
\usage{
getFeqTable(text)
}
\arguments{
\item{text}{plain text or a paragraph}
}
\value{
data frame of word and it's frequency.
}
\description{
getFeqTable
}
\examples{
\dontrun{
getFeqTable("India is Incredible!")
}
}
\author{
Jayachandra N
}
