% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_report.R
\name{getLibraryReport}
\alias{getLibraryReport}
\title{Get Library Report}
\usage{
getLibraryReport(packages)
}
\arguments{
\item{packages}{Vector of package names}
}
\value{
data.frame, status of required packages and their installation status
}
\description{
Get report on whether the given packages are installed on not
}
\details{
getLibraryReport
}
\examples{
getLibraryReport(c('dplyr', 'data.table'))
}
\author{
Jayachandra N
}
