% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cloud.R
\name{getWordCloud}
\alias{getWordCloud}
\title{getWordCloud}
\usage{
getWordCloud(d)
}
\arguments{
\item{d}{table of word's frequency}
}
\value{
Word cloud plot
}
\description{
getWordCloud
}
\examples{
\dontrun{
x <- getFeqTable("Hello! R is Great")
getWordCloud(x)
}
}
