% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_and_summarize.R
\name{groupByandSumarize}
\alias{groupByandSumarize}
\title{Group By And Summarize}
\usage{
groupByandSumarize(df, grp_col, summarise_col, FUN = mean)
}
\arguments{
\item{df}{data frame}

\item{grp_col}{column name to group}

\item{summarise_col}{column name to summarize}

\item{FUN}{function to summarize}
}
\value{
summarized table
}
\description{
Group by columns and summarize given data.
}
\details{
groupByandSumarize
}
\examples{
groupByandSumarize(mtcars, grp_col = c("am"), summarise_col = "hp",
FUN = "mean")
}
\author{
Jayachandra N
}
