% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_count.R
\name{missing_count}
\alias{missing_count}
\title{missing_count}
\usage{
missing_count(x)
}
\arguments{
\item{x}{vector}
}
\value{
Number of missing values in the given set of values
}
\description{
missing_count
}
\examples{
\dontrun{
missing_count(c(1,2,3))
missing_count(c(NA, 1, NA, "NULL", ""))
}
}
\author{
Jayachandra N
}
