% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyR.R
\name{shineMe}
\alias{shineMe}
\title{shineMe}
\usage{
shineMe()
}
\value{
shiny UI page
}
\description{
An R shiny app for shinyr UI.
}
\details{
shineMe
}
\examples{
\dontrun{
shineMe()
}
}
\author{
Jayachandra N
}
